Created a shared directory and moved baseResolver and pagination into it.
