+++ src/schema/ailments/ailment.args.ts
@@ -1,5 +1,5 @@
 import { ArgsType, Field } from 'type-graphql';
-import { PaginationArgs } from '../pagination/pagination.args';
+import { PaginationArgs } from '../shared';
 
 @ArgsType()
 export class AilmentArgs extends PaginationArgs {
