+++ src/schema/items/item.args.ts
@@ -1,6 +1,6 @@
 import { Length } from 'class-validator';
 import { ArgsType, Field, Int } from 'type-graphql';
-import { PaginationArgs } from '../pagination/pagination.args';
+import { PaginationArgs } from '../shared';
 
 @ArgsType()
 export class ItemArgs extends PaginationArgs {
