+++ src/schema/shared/index.ts
@@ -1,0 +1,4 @@
+import { createBaseResolver } from './baseResolver';
+import { PaginationArgs } from './pagination.args';
+
+export { createBaseResolver, PaginationArgs };
